Refactor for 2.0
