+++ azurerm/internal/services/web/parse/app_service_environment_test.go
@@ -1,4 +1,4 @@
-package web
+package parse
 
 import "testing"
 
@@ -46,7 +46,7 @@ func TestParseAppServiceEnvironmentID(t *testing.T) {
 	for _, v := range testData {
 		t.Logf("[DEBUG] Testing %q", v.Name)
 
-		actual, err := ParseAppServiceEnvironmentID(v.Input)
+		actual, err := AppServiceEnvironmentID(v.Input)
 		if err != nil {
 			if v.Expected == nil {
 				continue
@@ -61,43 +61,6 @@ func TestParseAppServiceEnvironmentID(t *testing.T) {
 
 		if actual.ResourceGroup != v.Expected.ResourceGroup {
 			t.Fatalf("Expected %q but got %q for Resource Group", v.Expected.ResourceGroup, actual.ResourceGroup)
-		}
-	}
-}
-
-func TestValidateAppServiceEnvironmentID(t *testing.T) {
-	cases := []struct {
-		ID    string
-		Valid bool
-	}{
-		{
-			ID:    "",
-			Valid: false,
-		},
-		{
-			ID:    "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/",
-			Valid: false,
-		},
-		{
-			ID:    "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/mygroup1/",
-			Valid: false,
-		},
-		{
-			ID:    "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/foo/providers/Microsoft.Web/hostingEnvironments/",
-			Valid: false,
-		},
-		{
-			ID:    "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/testGroup1/providers/Microsoft.Web/hostingEnvironments/TestASEv2",
-			Valid: true,
-		},
-	}
-	for _, tc := range cases {
-		t.Logf("[DEBUG] Testing Value %s", tc.ID)
-		_, errors := ValidateAppServiceEnvironmentID(tc.ID, "test")
-		valid := len(errors) == 0
-
-		if tc.Valid != valid {
-			t.Fatalf("Expected %t but got %t", tc.Valid, valid)
 		}
 	}
 }
