+++ azurerm/internal/services/web/validate/app_service_environment.go
@@ -1,0 +1,47 @@
+package validate
+
+import (
+	"fmt"
+	"regexp"
+
+	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/web/parse"
+)
+
+func AppServiceEnvironmentID(i interface{}, k string) (warnings []string, errors []error) {
+	v, ok := i.(string)
+	if !ok {
+		errors = append(errors, fmt.Errorf("expected type of %q to be string", k))
+		return
+	}
+
+	if _, err := parse.AppServiceEnvironmentID(v); err != nil {
+		errors = append(errors, fmt.Errorf("Can not parse %q as a resource id: %v", k, err))
+		return
+	}
+
+	return warnings, errors
+}
+
+func AppServiceEnvironmentName(v interface{}, k string) (warnings []string, errors []error) {
+	value := v.(string)
+
+	if matched := regexp.MustCompile(`^[0-9a-zA-Z][-0-9a-zA-Z]{0,61}[0-9a-zA-Z]$`).Match([]byte(value)); !matched {
+		errors = append(errors, fmt.Errorf("%q may only contain alphanumeric characters and dashes up to 60 characters in length, and must start and end in an alphanumeric", k))
+	}
+
+	return warnings, errors
+}
+
+func AppServiceEnvironmentPricingTier(v interface{}, k string) (warnings []string, errors []error) {
+	tier := v.(string)
+
+	valid := []string{"I1", "I2", "I3"}
+
+	for _, val := range valid {
+		if val == tier {
+			return
+		}
+	}
+	errors = append(errors, fmt.Errorf("pricing_tier must be one of %q", valid))
+	return warnings, errors
+}
