+++ azurerm/internal/services/web/validate/app_service_environment_test.go
@@ -1,0 +1,40 @@
+package validate
+
+import "testing"
+
+func TestValidateAppServiceEnvironmentID(t *testing.T) {
+	cases := []struct {
+		ID    string
+		Valid bool
+	}{
+		{
+			ID:    "",
+			Valid: false,
+		},
+		{
+			ID:    "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/",
+			Valid: false,
+		},
+		{
+			ID:    "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/mygroup1/",
+			Valid: false,
+		},
+		{
+			ID:    "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/foo/providers/Microsoft.Web/hostingEnvironments/",
+			Valid: false,
+		},
+		{
+			ID:    "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/testGroup1/providers/Microsoft.Web/hostingEnvironments/TestASEv2",
+			Valid: true,
+		},
+	}
+	for _, tc := range cases {
+		t.Logf("[DEBUG] Testing Value %s", tc.ID)
+		_, errors := AppServiceEnvironmentID(tc.ID, "test")
+		valid := len(errors) == 0
+
+		if tc.Valid != valid {
+			t.Fatalf("Expected %t but got %t", tc.Valid, valid)
+		}
+	}
+}
